legal-api/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
langchain-mongodb
langgraph
langgraph-checkpoint-mongodb
pydantic
pydantic-settings
opik
deepeval
sentence-transformers
numpy<2.0.0
pymongo
python-dotenv
loguru
motor
pypdf
